Moved serializer for floating point range to separate module
